IRkernel {IRkernel}	R Documentation

An R kernel for Jupyter.

Description

Jupyter speaks a JSON+ZMQ protocol to a 'kernel' which is responsible for executing code. This package is a kernel for the R language.

Usage

jupyter_option_defaults

Format

An object of class list of length 6.

Options

The following can be set/read via options(opt.name = ...) / getOption('opt.name')

jupyter.log_level

1L (errors), 2L (warnings), or 3L (debug). 1L is the default.

jupyter.pager_classes

Classes to use the pager for instead of displaying them inline. Default: help pages

jupyter.in_kernel

TRUE if this code is executed in a running kernel. Set to pretend being/not being in a kernel

jupyter.rich_display

Use more than just text display

jupyter.display_mimetypes

The formats emitted when any return value is to be displayed (default: all mimetypes listed here)

jupyter.plot_mimetypes

The plot formats emitted to the frontend when a plot is displayed. (default: image/png, application/pdf, and image/svg+xml)

See Also

installspec

[Package IRkernel version 0.8.12.9000 Index]