installspec {IRkernel}	R Documentation

Install the kernelspec to tell Jupyter about IRkernel.

Description

This can be called multiple times for different R interpreter, but you have to give a different name (and displayname to see a difference in the notebook UI). If the same name is give, it will overwrite older versions of the kernel spec with that name!

Usage

installspec(user = NULL, name = "ir", displayname = "R",
  rprofile = NULL, prefix = NULL)

Arguments

user
Install into user directory ($XDG_DATA_HOME/jupyter/kernels) or globally? (default: NULL but treated as TRUE if "prefix" is not specified)

name
The name of the kernel (default "ir")

displayname
The name which is displayed in the notebook (default: "R")

rprofile
(optional) Path to kernel-specific Rprofile (defaults to system-level settings)

prefix
(optional) Path to alternate directory to install kernelspec into (default: NULL)

Value

Exit code of the jupyter kernelspec install call.

[Package IRkernel version 0.8.12.9000 Index]