log_* {IRkernel}	R Documentation

Kernel logging functions

Description

A set of exported logging utilities that have the capability to be used in upstream projects. Log level and log file can be set via R package options e.g. options(jupyter.log_level = 2L) or from the environment variables JUPYTER_LOG_LEVEL and JUPYTER_LOGFILE.

Usage

log_debug(...)

log_info(...)

log_error(...)

Arguments

...
message to log

[Package IRkernel version 0.8.12.9000 Index]